+++ src/main/java/nl/hu/prbed/airline/customer/presentation/dto/CustomerDTO.java
@@ -9,7 +9,7 @@ public class CustomerDTO {
     public String firstName;
     public String lastName;
     public LocalDateTime dateOfBirth;
-    public Integer phoneNumber;
+    public Long phoneNumber;
     public String emailAddress;
     public String nationality;
