+++ src/main/java/nl/hu/prbed/airline/customer/presentation/dto/PassengerRequestDTO.java
@@ -10,7 +10,7 @@ public class PassengerRequestDTO {
     public String firstName;
     public String lastName;
     public LocalDateTime dateOfBirth;
-    public Integer phoneNumber;
+    public Long phoneNumber;
     public String emailAddress;
     public String nationality;
 }
